refactor: split script utils into multiple files
