+++ scripts/utils/index.py
@@ -1,57 +1,9 @@
-from typing import Mapping, MutableMapping, overload, Iterator, Literal, Iterable, TypeVar, TypedDict
-from datetime import datetime, timezone
-from requests.structures import CaseInsensitiveDict
-import itertools
-import re
+import os
 import json
 import logging
-import os
-import subprocess
-
-T = TypeVar('T')
-
-def configure_logging(verbosity):
-  if verbosity == 0:
-    level = logging.CRITICAL
-  elif verbosity == 1:
-    level = logging.INFO
-  else:
-    level = logging.DEBUG
-  logging.basicConfig(level=level, format='%(levelname)s: %(message)s')
-
-# from https://antonz.org/page-iterator/
-def paginate(iterable: 'Iterable[T]', page_size: int) -> 'Iterator[list[T]]':
-  it = iter(iterable)
-  slicer = lambda: list(itertools.islice(it, page_size))
-  return iter(slicer, [])
-
-#---
-# Processes
-#---
-
-class CommandError(RuntimeError):
-  pass
-
-def run_cmd(*args: str, allow_failure=False):
-  logging.info(f'> {" ".join(args)}')
-  rc = subprocess.run(args).returncode
-  if not allow_failure and rc != 0:
-    raise CommandError(f'external command exited with code {rc}')
-  return rc
-
-def capture_cmd(*args: str) -> bytes:
-  logging.debug(f'> {" ".join(args)}')
-  child = subprocess.run(args, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-  if child.returncode != 0:
-    raise CommandError(child.stderr.decode().strip())
-  elif len(child.stderr) > 0:
-    logging.error(child.stderr.decode())
-  return child.stdout
-
-
-#---
-# Index
-#---
+from typing import Mapping, MutableMapping, overload, Literal, Iterable, TypedDict, TypeVar
+from requests.structures import CaseInsensitiveDict
+from utils.core import *
 
 class Source(TypedDict):
   type: str
@@ -236,95 +188,3 @@ def write_aliases(index_dir: str, aliases: 'Mapping[str, Package]'):
 def write_index(index_dir: str, pkgs: 'Iterable[Package]', aliases: 'Mapping[str, Package]'):
   write_packages(index_dir, pkgs)
   write_aliases(index_dir, aliases)
-
-#---
-# Toolchains
-#---
-
-DEFAULT_ORIGIN = 'leanprover/lean4'
-
-class Release(TypedDict):
-  tag_name: str
-  published_at: str
-  html_url: str
-  prerelease: bool
-
-def query_releases(repo=DEFAULT_ORIGIN, paginate=True) -> 'Iterator[Release]':
-  out = capture_cmd(
-    'gh', 'api',
-    '--cache', '1h',
-    f'repos/{repo}/releases',
-    *(['--paginate'] if paginate else []),
-    '-q', '.[]'
-  )
-  return map(json.loads, out.decode().splitlines())
-
-class Toolchain(TypedDict):
-  name: str
-  version: int | None
-  tag: str
-  date: str
-  releaseUrl: str
-  prerelease: bool
-
-def utc_iso_now():
-  return datetime.now(timezone.utc).strftime("%Y-%m-%dT%H:%M:%SZ")
-
-def of_utc_iso(iso: str) -> datetime:
-  return datetime.strptime(iso, "%Y-%m-%dT%H:%M:%SZ").replace(tzinfo=timezone.utc)
-
-def toolchain_sort_key(t: Toolchain):
-    return (t['version'] or 0, of_utc_iso(t['date']))
-
-MIN_TOOLCHAIN_SORT_KEY = (0, datetime.min.replace(tzinfo=timezone.utc))
-
-TOOLCHAIN_VER_PATTERN = re.compile("v4\\.(\\d+)\\..*")
-def query_toolchains(repo=DEFAULT_ORIGIN) -> 'list[Toolchain]':
-  def toolchain_of_release(rel: Release) -> Toolchain:
-    match = TOOLCHAIN_VER_PATTERN.search(rel['tag_name'])
-    version = int(match.group(1)) if match is not None else None
-    return {
-      "name": f"{repo}:{rel['tag_name']}",
-      "version": version,
-      "tag": rel['tag_name'],
-      "date": rel['published_at'],
-      "releaseUrl": rel['html_url'],
-      "prerelease": rel['prerelease']
-    }
-  toolchains = map(toolchain_of_release, query_releases(repo))
-  return sorted(toolchains, key=toolchain_sort_key, reverse=True)
-
-def normalize_toolchain(toolchain: str):
-  parts = toolchain.split(':')
-  if len(parts) < 2:
-    origin = DEFAULT_ORIGIN
-    ver = parts[0]
-  else:
-    origin = parts[0]
-    ver = parts[1]
-  if ver[0].isdecimal():
-    ver = f'v{ver}'
-  return f'{origin}:{ver}'
-
-NIGHTLY_REPO='leanprover/lean4-nightly'
-def resolve_toolchain(toolchain: str):
-  toolchain = toolchain.strip()
-  if len(toolchain) == 0 or toolchain == 'package':
-    return None
-  elif toolchain == 'stable':
-    releases = filter(lambda r: not r['prerelease'], query_releases())
-    return f"{DEFAULT_ORIGIN}:{next(releases)['tag_name']}"
-  elif toolchain == 'nightly':
-    releases = query_releases(NIGHTLY_REPO, paginate=False)
-    return f"{DEFAULT_ORIGIN}:{next(releases)['tag_name']}"
-  elif toolchain == 'latest':
-    releases = query_releases(paginate=False)
-    return f"{DEFAULT_ORIGIN}:{next(releases)['tag_name']}"
-  else:
-    return normalize_toolchain(toolchain)
-
-def resolve_toolchains(toolchains: 'list[str]') -> 'set[str | None]':
-  if len(toolchains) == 0:
-    return set([None])
-  else:
-    return set(resolve_toolchain(t) for ts in toolchains for t in ts.split(','))
